GCP-465 & GCP-581
git-svn-id: https://efficio.atlassian.net/svn/GCP/trunk/IBPMiddleware@1457 4ebb2e00-0138-4608-bc6e-edf3948d4863

+++ src/main/java/org/generationcp/middleware/dao/ProjectMethodDAO.java
@@ -95,7 +95,7 @@ public class ProjectMethodDAO extends GenericDAO<ProjectMethod, Integer>{
             for (Object o : results) {
                 Object[] result = (Object[]) o;
                 if (result != null) {
-                    Long projectMethodId = Long.valueOf((Integer) result[0]);
+                    Integer projectMethodId = (Integer) result[0];
                     Integer methodId = (Integer) result[2];
                     ProjectMethod projectMethod = new ProjectMethod(projectMethodId, project, methodId);
                     toReturn.add(projectMethod);
